Updated index.js to render landing page, added landing page
