+++ src/components/pages/Landing/Landing.js
@@ -1,0 +1,18 @@
+import React from 'react';
+import 'antd/dist/antd.css';
+import { Card } from 'antd';
+import { Button } from 'antd';
+
+const Landing = () => {
+  return (
+    <>
+      <Card title="Requires your consent" style={{ width: 300 }}>
+        <Button type="primary">Take me away</Button>
+        <></>
+        <></>
+        <Button type="primary">Let me see!</Button>
+      </Card>
+    </>
+  );
+};
+export default Landing;
