+++ src/components/pages/Landing/index.js
@@ -1,0 +1,1 @@
+export { default as Landing } from './Landing';
